fix: types moved to each folders
