+++ pkg/passenger/passenger.go
@@ -5,16 +5,6 @@ import (
 	"time"
 )
 
-// Passenger stores information about a user.
-type Passenger struct {
-	ID        string    `json:"id"`
-	FirstName string    `json:"first_name"`
-	LastName  string    `json:"last_name"`
-	Email     string    `json:"email"`
-	Password  string    `json:"-"`
-	CreatedAt time.Time `json:"created_at"`
-}
-
 // PassengerService interface implements methods for managing passengers.
 type PassengerService interface {
 	GetPassengers() ([]Passenger, error)
